Use generated linux bindings

+++ rusl/src/unistd/close.rs
@@ -8,7 +8,7 @@ use crate::platform::Fd;
 /// See above
 #[inline]
 pub fn close(fd: Fd) -> crate::Result<()> {
-    let res = unsafe { syscall!(CLOSE, fd) } as i32;
+    let res = unsafe { syscall!(CLOSE, fd) };
     bail_on_below_zero!(res, "Failed `CLOSE` syscall");
     Ok(())
 }
