+++ rusl/src/unistd/write.rs
@@ -8,7 +8,7 @@ use crate::platform::Fd;
 /// See above for possible errors
 #[inline]
 pub fn write(fd: Fd, buf: &[u8]) -> crate::Result<usize> {
-    let res = unsafe { syscall!(WRITE, fd, buf.as_ptr(), buf.len()) as i32 };
+    let res = unsafe { syscall!(WRITE, fd, buf.as_ptr(), buf.len()) };
     bail_on_below_zero!(res, "`WRITE` syscall failed");
-    Ok(res as usize)
+    Ok(res)
 }
